refactor and add docs
