+++ arches_orm/models.py
@@ -1,4 +1,4 @@
 # Dummy module, to house dynamically generated
 # resource models.
-from .wkrm import resource_models
-globals().update(resource_models)
+from .wkrm import get_resource_models_for_adapter
+globals().update(get_resource_models_for_adapter()["by-class"])
