+++ arches_orm/utils.py
@@ -1,22 +1,23 @@
-from arches.app.models.resource import Resource
 import logging
-from .wkrm import resource_models, resource_models_by_graph_id
-from .wrapper import ResourceModelWrapper
+from .adapter import get_adapter
+from .wkrm import get_resource_models_for_adapter
 
 logger = logging.getLogger(__name__)
 
 
-def get_well_known_resource_model_by_class_name(class_name, default=None):
+def get_well_known_resource_model_by_class_name(class_name, default=None, adapter: str | None=None):
     """Turns a class-name as a string into a well-known resource model wrapper."""
+    resource_models = get_resource_models_for_adapter(adapter)["by-class"]
     return resource_models.get(class_name, default)
 
 
-def get_well_known_resource_model_by_graph_id(graphid, default=None):
+def get_well_known_resource_model_by_graph_id(graphid, default=None, adapter: str | None=None):
     """Turns a graph into a well-known resource model wrapper, by ID, if known."""
+    resource_models_by_graph_id = get_resource_models_for_adapter(adapter)["by-graph-id"]
     return resource_models_by_graph_id.get(str(graphid), default)
 
 
-def attempt_well_known_resource_model(resource_id, from_prefetch=None, **kwargs):
+def attempt_well_known_resource_model(resource_id, from_prefetch=None, adapter=None, **kwargs):
     """Attempts to find and create a well-known resource from a resource ID
 
     This is the simplest entry-point if you do not know the model of the resource
@@ -24,17 +25,7 @@ def attempt_well_known_resource_model(resource_id, from_prefetch=None, **kwargs)
     is not matched.
     """
 
-    resource = (
-        from_prefetch(resource_id)
-        if from_prefetch is not None
-        else Resource.objects.get(pk=resource_id)
+    return get_adapter(adapter).load_from_id(
+        resource_id=resource_id,
+        from_prefetch=from_prefetch
     )
-    if resource is None:
-        logger.error("Tried to load non-existent WKRM: %s", resource_id)
-        return None
-    if isinstance(resource, ResourceModelWrapper):
-        return resource
-    wkrm = get_well_known_resource_model_by_graph_id(resource.graph_id, default=None)
-    if wkrm:
-        return wkrm.from_resource(resource, **kwargs)
-    return None
