+++ arches_orm/view_models/_base.py
@@ -1,0 +1,5 @@
+class WKRI:
+    ...
+
+class ViewModel:
+    _parent_pseudo_node = None
