+++ arches_orm/wkrm.py
@@ -1,7 +1,7 @@
 import logging
 from django.conf import settings
 from typing import Callable
-from .wrapper import get_adapter
+from .adapter import get_adapter
 
 
 logger = logging.getLogger(__name__)
@@ -36,11 +36,11 @@ WELL_KNOWN_RESOURCE_MODELS = [
 ]
 
 
-def _make_wkrm(wkrm_definition, adapter: str | None=None):
+def _make_wkrm(wkrm_definition, adapter):
     try:
         return type(
             wkrm_definition.model_class_name,
-            (get_adapter(adapter).get_wrapper(),),
+            (adapter.get_wrapper(),),
             {},
             well_known_resource_model=wkrm_definition,
         )
@@ -52,7 +52,17 @@ def _make_wkrm(wkrm_definition, adapter: str | None=None):
         logger.exception(e)
 
 
-resource_models = {
-    wkrm.model_class_name: _make_wkrm(wkrm) for wkrm in WELL_KNOWN_RESOURCE_MODELS
-}
-resource_models_by_graph_id = {rm.graphid: rm for rm in resource_models.values() if rm}
+resource_models = {}
+def get_resource_models_for_adapter(adapter_name: str | None=None):
+    adapter = get_adapter(adapter_name)
+    if str(adapter) not in resource_models:
+        resource_models[str(adapter)] = {}
+        resource_models[str(adapter)]["by-class"] = {
+            wkrm.model_class_name: _make_wkrm(wkrm, adapter) for wkrm in WELL_KNOWN_RESOURCE_MODELS
+        }
+        resource_models[str(adapter)]["by-graph-id"] = {
+            rm.graphid: rm for rm in resource_models[str(adapter)]["by-class"].values() if rm
+        }
+    return resource_models[str(adapter)]
+
+get_resource_models_for_adapter()
