+++ docs/arches_orm/hooks.md
@@ -1,0 +1,12 @@
+# Module arches_orm.hooks
+
+??? example "View Source"
+        from .wrapper import get_adapter
+
+        HOOKS = get_adapter().get_hooks()
+
+## Variables
+
+```python3
+HOOKS
+```
