+++ docs/arches_orm/models.md
@@ -1,0 +1,16 @@
+# Module arches_orm.models
+
+??? example "View Source"
+        # Dummy module, to house dynamically generated
+
+        # resource models.
+
+        from .wkrm import get_resource_models_for_adapter
+
+        globals().update(get_resource_models_for_adapter()["by-class"])
+
+## Variables
+
+```python3
+Person
+```
